Rename Requests folder
