+++ api/app/Http/Requests/ApiFormRequest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types = 1);
 
-namespace App\Http\Request;
+namespace App\Http\Requests;
 
 use Illuminate\Foundation\Http\FormRequest;
