+++ api/app/Http/Requests/Auth/RegisterHttpRequest.php
@@ -2,9 +2,12 @@
 
 declare(strict_types = 1);
 
-namespace App\Http\Request\Api\Auth;
+namespace App\Http\Requests\Auth;
 
-use App\Http\Request\ApiFormRequest;
+use App\Http\Requests\ApiFormRequest;
+use Illuminate\Contracts\Validation\Validator;
+use Illuminate\Http\Exceptions\HttpResponseException;
+use Illuminate\Http\JsonResponse;
 
 final class RegisterHttpRequest extends ApiFormRequest
 {
@@ -19,7 +22,7 @@ final class RegisterHttpRequest extends ApiFormRequest
             'email' => 'required|email|unique:users|max:50',
             'name' => 'required|string|between:2,100',
             'password' => 'required|confirmed|min:8|string',
-            'timezone' => 'required'
+//            'timezone' => 'required'
         ];
     }
 }
